LayoutTests/imported/w3c/web-platform-tests/preload/preload-error.sub-expected.txt

CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Cannot load image http://127.0.0.1:8800/preload/resources/square.png?label=image due to access control checks.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8801/preload/resources/square.png?pipe=header(Access-Control-Allow-Origin,*) because it appears in neither the img-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Cannot load image http://localhost:8800/common/redirect.py?location=http%3A%2F%2F127.0.0.1%3A8801%2Fpreload%2Fresources%2Fsquare.png%3Fpipe%3Dheader%28Access-Control-Allow-Origin%2C*%29&label=image due to access control checks.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8801/preload/resources/square.png?pipe=header(Access-Control-Allow-Origin,*) because it appears in neither the img-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8801/preload/resources/dummy.css?pipe=header(Access-Control-Allow-Origin,*) because it appears in neither the style-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Refused to load http://127.0.0.1:8801/preload/resources/dummy.js?pipe=header(Access-Control-Allow-Origin,*) because it appears in neither the script-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/preload/resources/dummy.xml?label=xhr due to access control checks.
CONSOLE MESSAGE: Refused to connect to http://127.0.0.1:8801/preload/resources/dummy.xml?pipe=header(Access-Control-Allow-Origin,*) because it appears in neither the connect-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Blocked by Content Security Policy.
CONSOLE MESSAGE: Refused to connect to http://127.0.0.1:8801/preload/resources/dummy.xml?pipe=header(Access-Control-Allow-Origin,*) because it appears in neither the connect-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Blocked by Content Security Policy.
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8801/preload/resources/dummy.xml?pipe=header(Access-Control-Allow-Origin,*) due to access control checks.
CONSOLE MESSAGE: Origin http://localhost:8800 is not allowed by Access-Control-Allow-Origin. Status code: 200
CONSOLE MESSAGE: Fetch API cannot load http://127.0.0.1:8800/preload/resources/dummy.xml?label=fetch due to access control checks.
CONSOLE MESSAGE: Refused to connect to http://127.0.0.1:8801/preload/resources/dummy.xml?pipe=header(Access-Control-Allow-Origin,*) because it appears in neither the connect-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Blocked by Content Security Policy.
CONSOLE MESSAGE: Refused to connect to http://127.0.0.1:8801/preload/resources/dummy.xml?pipe=header(Access-Control-Allow-Origin,*) because it appears in neither the connect-src directive nor the default-src directive of the Content Security Policy.
CONSOLE MESSAGE: Blocked by Content Security Policy.
CONSOLE MESSAGE: Fetch API cannot load http://127.0.0.1:8801/preload/resources/dummy.xml?pipe=header(Access-Control-Allow-Origin,*) due to access control checks.
CONSOLE MESSAGE: SyntaxError: Unexpected token '<'
CONSOLE MESSAGE: Did not parse stylesheet at 'http://localhost:8800/preload/resources/dummy.js?pipe=header%28Content-Type%2Cnot%2Fcss%29&label=style' because non CSS MIME types are not allowed in strict mode.
CONSOLE MESSAGE: Refused to execute http://localhost:8800/preload/resources/dummy.css?pipe=header%28Content-Type%2Cimage%2Fnot-javascript%29&label=script as script because image/not-javascript is not a script MIME type.
CONSOLE MESSAGE: Did not parse stylesheet at 'http://localhost:8800/preload/resources/dummy.js?pipe=header%28Content-Type%2Cnot%2Fcss%29%7Cheader%28X-Content-Type-Options%2Cnosniff%29&label=style' because non CSS MIME types are not allowed in strict mode.
CONSOLE MESSAGE: Refused to execute http://localhost:8800/preload/resources/dummy.css?pipe=header%28Content-Type%2Ctext%2Fnotjavascript%29%7Cheader%28X-Content-Type-Options%2Cnosniff%29&label=script as script because "X-Content-Type-Options: nosniff" was given and its Content-Type is not a script MIME type.

PASS success (image): preload events
FAIL success (image): main assert_less_than_equal: http://localhost:8800/preload/resources/square.png?label=image should be downloaded from network at most once expected a number less than or equal to 1 but got 2
PASS 404 (image): preload events
FAIL 404 (image): main assert_less_than_equal: http://localhost:8800/preload/resources/square.png?pipe=status%28404%29&label=image should be downloaded from network at most once expected a number less than or equal to 1 but got 2
PASS CORS (image): preload events
PASS CORS (image): main
PASS CORS-error (image): preload events
FAIL CORS-error (image): main assert_greater_than: http://127.0.0.1:8800/preload/resources/square.png?label=image should be loaded expected a number greater than 0 but got 0
PASS CSP-error (image): preload events
FAIL CSP-error (image): main assert_greater_than: http://localhost:8800/common/redirect.py?location=http%3A%2F%2F127.0.0.1%3A8801%2Fpreload%2Fresources%2Fsquare.png%3Fpipe%3Dheader%28Access-Control-Allow-Origin%2C*%29&label=image should be loaded expected a number greater than 0 but got 0
PASS success (style): preload events
PASS success (style): main
FAIL 404 (style): preload events assert_unreached: preload onerror Reached unreachable code
PASS 404 (style): main
PASS CORS (style): preload events
PASS CORS (style): main
PASS CORS-error (style): preload events
FAIL CORS-error (style): main assert_greater_than: http://127.0.0.1:8800/preload/resources/dummy.css?label=style should be loaded expected a number greater than 0 but got 0
PASS CSP-error (style): preload events
FAIL CSP-error (style): main assert_greater_than: http://localhost:8800/common/redirect.py?location=http%3A%2F%2F127.0.0.1%3A8801%2Fpreload%2Fresources%2Fdummy.css%3Fpipe%3Dheader%28Access-Control-Allow-Origin%2C*%29&label=style should be loaded expected a number greater than 0 but got 0
PASS success (script): preload events
PASS success (script): main
FAIL 404 (script): preload events assert_unreached: preload onerror Reached unreachable code
PASS 404 (script): main
PASS CORS (script): preload events
PASS CORS (script): main
PASS CORS-error (script): preload events
FAIL CORS-error (script): main assert_greater_than: http://127.0.0.1:8800/preload/resources/dummy.js?label=script should be loaded expected a number greater than 0 but got 0
PASS CSP-error (script): preload events
FAIL CSP-error (script): main assert_greater_than: http://localhost:8800/common/redirect.py?location=http%3A%2F%2F127.0.0.1%3A8801%2Fpreload%2Fresources%2Fdummy.js%3Fpipe%3Dheader%28Access-Control-Allow-Origin%2C*%29&label=script should be loaded expected a number greater than 0 but got 0
PASS success (xhr): preload events
FAIL success (xhr): main assert_less_than_equal: http://localhost:8800/preload/resources/dummy.xml?label=xhr should be downloaded from network at most once expected a number less than or equal to 1 but got 2
PASS 404 (xhr): preload events
FAIL 404 (xhr): main assert_less_than_equal: http://localhost:8800/preload/resources/dummy.xml?pipe=status%28404%29&label=xhr should be downloaded from network at most once expected a number less than or equal to 1 but got 2
PASS CORS (xhr): preload events
PASS CORS (xhr): main
PASS CORS-error (xhr): preload events
FAIL CORS-error (xhr): main assert_greater_than: http://127.0.0.1:8800/preload/resources/dummy.xml?label=xhr should be loaded expected a number greater than 0 but got 0
PASS CSP-error (xhr): preload events
FAIL CSP-error (xhr): main assert_greater_than: http://localhost:8800/common/redirect.py?location=http%3A%2F%2F127.0.0.1%3A8801%2Fpreload%2Fresources%2Fdummy.xml%3Fpipe%3Dheader%28Access-Control-Allow-Origin%2C*%29&label=xhr should be loaded expected a number greater than 0 but got 0
PASS success (fetch): preload events
FAIL success (fetch): main assert_less_than_equal: http://localhost:8800/preload/resources/dummy.xml?label=fetch should be downloaded from network at most once expected a number less than or equal to 1 but got 2
PASS 404 (fetch): preload events
FAIL 404 (fetch): main assert_less_than_equal: http://localhost:8800/preload/resources/dummy.xml?pipe=status%28404%29&label=fetch should be downloaded from network at most once expected a number less than or equal to 1 but got 2
PASS CORS (fetch): preload events
PASS CORS (fetch): main
PASS CORS-error (fetch): preload events
FAIL CORS-error (fetch): main assert_greater_than: http://127.0.0.1:8800/preload/resources/dummy.xml?label=fetch should be loaded expected a number greater than 0 but got 0
PASS CSP-error (fetch): preload events
FAIL CSP-error (fetch): main assert_greater_than: http://localhost:8800/common/redirect.py?location=http%3A%2F%2F127.0.0.1%3A8801%2Fpreload%2Fresources%2Fdummy.xml%3Fpipe%3Dheader%28Access-Control-Allow-Origin%2C*%29&label=fetch should be loaded expected a number greater than 0 but got 0
FAIL Decode-error (image): preload events assert_unreached: preload onerror Reached unreachable code
PASS Decode-error (image): main
PASS Decode-error (style): preload events
PASS Decode-error (style): main
PASS Decode-error (script): preload events
PASS Decode-error (script): main
PASS MIME-error (image): preload events
FAIL MIME-error (image): main assert_less_than_equal: http://localhost:8800/preload/resources/square.png?pipe=header%28Content-Type%2Ctext%2Fnotimage%29&label=image should be downloaded from network at most once expected a number less than or equal to 1 but got 2
PASS MIME-error (script): preload events
PASS MIME-error (script): main
PASS MIME-error (style): preload events
PASS MIME-error (style): main
FAIL MIME-blocked (script): preload events assert_unreached: preload onload Reached unreachable code
PASS MIME-blocked (script): main
FAIL MIME-blocked-nosniff (style): preload events assert_unreached: preload onload Reached unreachable code
PASS MIME-blocked-nosniff (style): main
FAIL MIME-blocked-nosniff (script): preload events assert_unreached: preload onload Reached unreachable code
PASS MIME-blocked-nosniff (script): main
